Move connect testcase into tests folder
